+++ examples/tutorial/mantine/src/pages/posts/list.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { useTable, ColumnDef, flexRender } from "@pankod/refine-react-table";
+import { GetManyResponse, useMany } from "@pankod/refine-core";
 import {
     Box,
     Group,
@@ -14,7 +15,7 @@ import {
 } from "@pankod/refine-mantine";
 
 import { ColumnFilter, ColumnSorter } from "../../components/table";
-import { FilterElementProps, IPost } from "../../interfaces";
+import { FilterElementProps, ICategory, IPost } from "../../interfaces";
 
 export const PostList: React.FC = () => {
     const columns = React.useMemo<ColumnDef<IPost>[]>(
@@ -51,6 +52,21 @@ export const PostList: React.FC = () => {
                         );
                     },
                     filterOperator: "eq",
+                },
+            },
+            {
+                id: "category.id",
+                header: "Category",
+                enableColumnFilter: false,
+                accessorKey: "category.id",
+                cell: function render({ getValue, table }) {
+                    const meta = table.options.meta as {
+                        categoriesData: GetManyResponse<ICategory>;
+                    };
+                    const category = meta.categoriesData?.data.find(
+                        (item) => item.id === getValue(),
+                    );
+                    return category?.title ?? "Loading...";
                 },
             },
             {
@@ -92,10 +108,33 @@ export const PostList: React.FC = () => {
     const {
         getHeaderGroups,
         getRowModel,
-        refineCore: { setCurrent, pageCount, current },
+        setOptions,
+        refineCore: {
+            setCurrent,
+            pageCount,
+            current,
+            tableQueryResult: { data: tableData },
+        },
     } = useTable({
         columns,
     });
+
+    const categoryIds = tableData?.data?.map((item) => item.category.id) ?? [];
+    const { data: categoriesData } = useMany<ICategory>({
+        resource: "categories",
+        ids: categoryIds,
+        queryOptions: {
+            enabled: categoryIds.length > 0,
+        },
+    });
+
+    setOptions((prev) => ({
+        ...prev,
+        meta: {
+            ...prev.meta,
+            categoriesData,
+        },
+    }));
 
     return (
         <ScrollArea>